requirements.txt
dbt-core==1.7.*
dbt-postgres==1.7.*
astronomer-cosmos
apache-airflow
pendulum